feat: create custom hook to separate grouping logics from component
