+++ react/hooks/index.ts
@@ -3,6 +3,7 @@ export * from './useClearCart'
 export * from './useFetchCustomerCredit'
 export * from './useFetchPrices'
 export * from './useFomatPrice'
+export * from './useGroupedProducts'
 export * from './useOrderFormCustom'
 export * from './useOrganization'
 export * from './usePermissions'
